+++ ontolutils/ex/ssn/__init__.py
@@ -1,12 +1,25 @@
-from .ssn import Property, Accuracy, SystemCapability, MeasurementRange, SystemProperty, Condition, System, __version__
+from .ssn import (Property, Accuracy, SystemCapability,
+                  MeasurementRange, SystemProperty, Condition,
+                  System, ObservableProperty, Observation,
+                  Actuator, Sensor, Sampler,
+                  Platform, FeatureOfInterest, Result,
+                  __version__)
 
 __all__ = [
     "Property",
-    "Accuracy",
-    "SystemCapability",
-    "MeasurementRange",
     "SystemProperty",
     "Condition",
+    "SystemCapability",
     "System",
+    "ObservableProperty",
+    "Actuator",
+    "Sensor",
+    "Sampler",
+    "Platform",
+    "FeatureOfInterest",
+    "Result",
+    "Observation",
+    "MeasurementRange",
+    "Accuracy",
     "__version__",
 ]